Pzem017 Library for Arduino
This Library is based on ModbusMaster by Doc Walker , as altered by  EvertDekker.com, and few changes by me , so it can be installed as separate library. Though I am not sure if it can be used simultaneously with ModbusMaster library or not. All credits to original authors, I have not committed to their code, as I am new at Github and though I may messup their code too.